fix: Divided mypage.js file by page and function
